+++ packages/data/src/namespace-store/index.js
@@ -155,9 +155,11 @@ function createReduxStore( key, options, registry ) {
  */
 function mapSelectors( selectors, store, registry ) {
 	const createStateSelector = ( registeredSelector ) => {
-		const selector = registeredSelector.isRegistrySelector ? registeredSelector( registry.select ) : registeredSelector;
+		const registrySelector = registeredSelector.isRegistrySelector ?
+			registeredSelector( registry.select ) :
+			registeredSelector;
 
-		return function runSelector() {
+		const selector = function runSelector() {
 			// This function is an optimized implementation of:
 			//
 			//   selector( store.getState(), ...arguments )
@@ -172,8 +174,10 @@ function mapSelectors( selectors, store, registry ) {
 				args[ i + 1 ] = arguments[ i ];
 			}
 
-			return selector( ...args );
+			return registrySelector( ...args );
 		};
+		selector.hasResolver = false;
+		return selector;
 	};
 
 	return mapValues( selectors, createStateSelector );
@@ -213,10 +217,11 @@ function mapResolvers( resolvers, selectors, store ) {
 	const mapSelector = ( selector, selectorName ) => {
 		const resolver = resolvers[ selectorName ];
 		if ( ! resolver ) {
+			selector.hasResolver = false;
 			return selector;
 		}
 
-		return ( ...args ) => {
+		const selectorResolver = ( ...args ) => {
 			async function fulfillSelector() {
 				const state = store.getState();
 				if ( typeof resolver.isFulfilled === 'function' && resolver.isFulfilled( state, ...args ) ) {
@@ -236,6 +241,8 @@ function mapResolvers( resolvers, selectors, store ) {
 			fulfillSelector( ...args );
 			return selector( ...args );
 		};
+		selectorResolver.hasResolver = true;
+		return selectorResolver;
 	};
 
 	return {