+++ packages/block-library/src/image/edit.native.js
@@ -2,7 +2,7 @@
  * External dependencies
  */
 import React from 'react';
-import { View, ImageBackground, Text, TouchableWithoutFeedback } from 'react-native';
+import { View, ImageBackground, Text, TouchableWithoutFeedback, Dimensions } from 'react-native';
 import {
 	requestMediaImport,
 	mediaUploadSync,
@@ -15,9 +15,9 @@ import { isEmpty } from 'lodash';
  * WordPress dependencies
  */
 import {
+	BottomSheet,
 	Toolbar,
 	ToolbarButton,
-	Dashicon,
 } from '@wordpress/components';
 import {
 	MediaPlaceholder,
@@ -26,7 +26,6 @@ import {
 	RichText,
 	BlockControls,
 	InspectorControls,
-	BottomSheet,
 } from '@wordpress/block-editor';
 import { __, sprintf } from '@wordpress/i18n';
 import { isURL } from '@wordpress/url';
@@ -37,9 +36,14 @@ import { doAction, hasAction } from '@wordpress/hooks';
  */
 import styles from './styles.scss';
 import MediaUploadProgress from './media-upload-progress';
+import SvgIcon from './icon';
+import SvgIconRetry from './icon-retry';
 
 const LINK_DESTINATION_CUSTOM = 'custom';
 const LINK_DESTINATION_NONE = 'none';
+
+// Default Image ratio 4:3
+const IMAGE_ASPECT_RATIO = 4 / 3;
 
 class ImageEdit extends React.Component {
 	constructor( props ) {
@@ -187,6 +191,14 @@ class ImageEdit extends React.Component {
 		}
 	}
 
+	getIcon( isRetryIcon ) {
+		if ( isRetryIcon ) {
+			return <SvgIconRetry fill={ styles.iconRetry.fill } />;
+		}
+
+		return <SvgIcon fill={ styles.icon.fill } />;
+	}
+
 	render() {
 		const { attributes, isSelected, setAttributes } = this.props;
 		const { url, caption, height, width, alt, href, id } = attributes;
@@ -199,22 +211,16 @@ class ImageEdit extends React.Component {
 			this.setState( { showSettings: false } );
 		};
 
-		const toolbarEditButton = (
-			<MediaUpload mediaType={ MEDIA_TYPE_IMAGE }
-				onSelectURL={ this.onSelectMediaUploadOption }
-				render={ ( { open, getMediaOptions } ) => {
-					return (
-						<Toolbar>
-							{ getMediaOptions() }
-							<ToolbarButton
-								title={ __( 'Edit image' ) }
-								icon="edit"
-								onClick={ open }
-							/>
-						</Toolbar>
-					);
-				} } >
-			</MediaUpload>
+		const getToolbarEditButton = ( open ) => (
+			<BlockControls>
+				<Toolbar>
+					<ToolbarButton
+						title={ __( 'Edit image' ) }
+						icon="edit"
+						onClick={ open }
+					/>
+				</Toolbar>
+			</BlockControls>
 		);
 
 		const getInspectorControls = () => (
@@ -231,6 +237,7 @@ class ImageEdit extends React.Component {
 					onChangeValue={ this.onSetLinkDestination }
 					autoCapitalize="none"
 					autoCorrect={ false }
+					keyboardType="url"
 				/>
 				<BottomSheet.Cell
 					icon={ 'editor-textcolor' }
@@ -255,15 +262,17 @@ class ImageEdit extends React.Component {
 					<MediaPlaceholder
 						mediaType={ MEDIA_TYPE_IMAGE }
 						onSelectURL={ this.onSelectMediaUploadOption }
+						icon={ this.getIcon( false ) }
+						onFocus={ this.props.onFocus }
 					/>
 				</View>
 			);
 		}
 
-		return (
+		const imageContainerHeight = Dimensions.get( 'window' ).width / IMAGE_ASPECT_RATIO;
+		const getImageComponent = ( openMediaOptions, getMediaOptions ) => (
 			<TouchableWithoutFeedback
 				accessible={ ! isSelected }
-
 				accessibilityLabel={ sprintf(
 					/* translators: accessibility text. 1: image alt text. 2: image caption. */
 					__( 'Image block. %1$s. %2$s' ),
@@ -272,14 +281,14 @@ class ImageEdit extends React.Component {
 				) }
 				accessibilityRole={ 'button' }
 				onPress={ this.onImagePressed }
+				onLongPress={ openMediaOptions }
 				disabled={ ! isSelected }
 			>
 				<View style={ { flex: 1 } }>
 					{ getInspectorControls() }
+					{ getMediaOptions() }
 					{ ( ! this.state.isCaptionSelected ) &&
-						<BlockControls>
-							{ toolbarEditButton }
-						</BlockControls>
+						getToolbarEditButton( openMediaOptions )
 					}
 					<InspectorControls>
 						<ToolbarButton
@@ -297,24 +306,36 @@ class ImageEdit extends React.Component {
 						onFinishMediaUploadWithSuccess={ this.finishMediaUploadWithSuccess }
 						onFinishMediaUploadWithFailure={ this.finishMediaUploadWithFailure }
 						onMediaUploadStateReset={ this.mediaUploadStateReset }
-						renderContent={ ( { isUploadInProgress, isUploadFailed, finalWidth, finalHeight, imageWidthWithinContainer, retryIconName, retryMessage } ) => {
+						renderContent={ ( { isUploadInProgress, isUploadFailed, finalWidth, finalHeight, imageWidthWithinContainer, retryMessage } ) => {
 							const opacity = isUploadInProgress ? 0.3 : 1;
+							const icon = this.getIcon( isUploadFailed );
+
+							const iconContainer = (
+								<View style={ styles.modalIcon }>
+									{ icon }
+								</View>
+							);
+
 							return (
 								<View style={ { flex: 1 } } >
-									{ ! imageWidthWithinContainer && <View style={ styles.imageContainer } >
-										<Dashicon icon={ 'format-image' } size={ 300 } />
-									</View> }
+									{ ! imageWidthWithinContainer &&
+										<View style={ [ styles.imageContainer, { height: imageContainerHeight } ] } >
+											{ this.getIcon( false ) }
+										</View> }
 									<ImageBackground
+										accessible={ true }
+										disabled={ ! isSelected }
+										accessibilityLabel={ alt }
+										accessibilityHint={ __( 'Double tap and hold to edit' ) }
+										accessibilityRole={ 'imagebutton' }
 										style={ { width: finalWidth, height: finalHeight, opacity } }
 										resizeMethod="scale"
 										source={ { uri: url } }
 										key={ url }
-										accessible={ true }
-										accessibilityLabel={ alt }
 									>
 										{ isUploadFailed &&
-											<View style={ styles.imageContainer } >
-												<Dashicon icon={ retryIconName } ariaPressed={ 'dashicon-active' } />
+											<View style={ [ styles.imageContainer, { flex: 1, backgroundColor: 'rgba(0, 0, 0, 0.5)' } ] } >
+												{ iconContainer }
 												<Text style={ styles.uploadFailedText }>{ retryMessage }</Text>
 											</View>
 										}
@@ -357,6 +378,15 @@ class ImageEdit extends React.Component {
 					) }
 				</View>
 			</TouchableWithoutFeedback>
+		);
+
+		return (
+			<MediaUpload mediaType={ MEDIA_TYPE_IMAGE }
+				onSelectURL={ this.onSelectMediaUploadOption }
+				render={ ( { open, getMediaOptions } ) => {
+					return getImageComponent( open, getMediaOptions );
+				} }
+			/>
 		);
 	}
 }