+++ packages/data/src/index.js
@@ -12,8 +12,15 @@ import * as plugins from './plugins';
 export { default as withSelect } from './components/with-select';
 export { default as withDispatch } from './components/with-dispatch';
 export { default as withRegistry } from './components/with-registry';
-export { default as RegistryProvider, RegistryConsumer } from './components/registry-provider';
-export { default as __experimentalAsyncModeProvider } from './components/async-mode-provider';
+export {
+	RegistryProvider,
+	RegistryConsumer,
+	useRegistry,
+} from './components/registry-provider';
+export { default as useSelect } from './components/use-select';
+export {
+	AsyncModeProvider as __experimentalAsyncModeProvider,
+} from './components/async-mode-provider';
 export { createRegistry } from './registry';
 export { plugins };
 export { createRegistrySelector, createRegistryControl } from './factory';