+++ packages/editor/CHANGELOG.md
@@ -8,6 +8,7 @@
 
 - Refactor setupEditor effects to action-generator using controls ([#14513](https://github.com/WordPress/gutenberg/pull/14513))
 - Remove redux-multi dependency (no longer needed/used with above refactor)
+- Replace internal controls definitions with usage of new @wordpress/data-controls package (see [#15435](https://github.com/WordPress/gutenberg/pull/15435)
 
 ## 9.1.0 (2019-03-06)